pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kd-gat"
version = "1.0.0"
description = "CAN Bus Intrusion Detection via Knowledge Distillation (VGAE → GAT → DQN)"
requires-python = ">=3.10,<3.13"
dependencies = [
    "torch>=2.8.0,<2.9",
    "torch-geometric>=2.7.0",
    "torch-scatter",
    "torch-sparse",
    "torch-cluster",
    "pytorch-lightning>=2.6.0",
    "numpy>=1.26.4",
    "pandas>=2.2.3",
    "pyarrow>=14.0.0",
    "pydantic>=2.10",
    "pyyaml>=6.0",
    "psutil>=6.0.0",
    "scikit-learn>=1.5.0",
    "wandb>=0.18",
    "ray[default]>=2.49",
    "ray[tune]>=2.49",
    "optuna>=4.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "boto3",
    "s3fs",
    "umap-learn>=0.5",
    "seaborn>=0.13",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "httpx>=0.27", "ruff>=0.4", "duckdb>=1.0.0"]
baselines = ["pygod>=1.1"]

[project.scripts]
graphids = "graphids.pipeline.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["graphids"]

[tool.uv.sources]
torch-scatter = { index = "pyg-cu126" }
torch-sparse = { index = "pyg-cu126" }
torch-cluster = { index = "pyg-cu126" }

[[tool.uv.index]]
name = "pyg-cu126"
url = "https://data.pyg.org/whl/torch-2.8.0+cu126.html"
explicit = true
format = "flat"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (>30s, deselect with '-m not slow')",
    "slurm: marks tests that require SLURM submission",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501", "E402", "E741", "W291", "W292", "W293", "F541", "B905", "B017"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["F401"]
"graphids/pipeline/stages/*" = ["F401"]
"graphids/pipeline/orchestration/*" = ["F401"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "graphids.config.*"
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "torch_geometric.*",
    "pytorch_lightning.*",
    "ray.*",
    "wandb.*",
    "optuna.*",
    "sklearn.*",
    "umap.*",
    "seaborn.*",
    "boto3.*",
    "s3fs.*",
]
ignore_missing_imports = true
